Add python {sum,log}2junit scripts
This adds a python based sum2junit which analyzes a sumfile and emits
a JUnit test log file corresponding to that run. This differs from the
shell variant in that it supports multiple test variantion in the same
sum file, and emits these as seperate testsuites, using the variation
as a class name in the testsuite. Additionally passing tests are also
printed so these can be tracked across runs.

This also adds a first version of a log2unit script which attempts to
parse a log file in step with a sumfile and extract relevant snippets
which are stored as the "system-out" element of each testcase. This
relies on the .sum and .log presenting results in the same order. A
future version of this script will support merging sum/log files such
that results can be taken directly from e.g. the GCC test suite partial
log files, and merged into one JUnit output.

ChangeLog:

	* contrib/log2junit.sh: Added.
	* contrib/sum2junit.sh: Added.